verbs: Move fi_getinfo related functionality into new file
